// src/schemas/objectWithRest/types.ts
import type { BaseIssue } from "../../types";

/**
 * Object with rest issue type.
 */
export interface ObjectWithRestIssue extends BaseIssue<unknown> {
	/**
	 * The issue kind.
	 */
	readonly kind: "schema";
	/**
	 * The issue type.
	 */
	readonly type: "object_with_rest";
	/**
	 * The expected property.
	 */
	readonly expected: "Object";
}